LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-004-expected.txt




PASS #target > div 1
FAIL #target > div 2 assert_equals:
<div style="grid-area: 1 / 2 / 4 / 2;" data-offset-y="100"><span></span><br><span></span></div>
offsetTop expected 100 but got 0
